+++ R/ArrayTimeBySpeciesBySize-class.R
@@ -254,37 +254,27 @@ ArrayTimeBySpeciesBySize_slice <- function(x, time = NULL) {
                        units = units, params = params)
 }
 
-#' @rdname plot
-#' @usage NULL
-#' @exportS3Method plotly::ggplotly
+#' @rdname plotHover
 #' @examples
 #' \donttest{
-#' ggplotly(getFMort(NS_sim))
+#' plotHover(getFMort(NS_sim))
 #' }
-ggplotly.ArrayTimeBySpeciesBySize <- function(p = ggplot2::last_plot(),
-                                              width = NULL, height = NULL,
-                                              tooltip = "all",
-                                              dynamicTicks = FALSE,
-                                              layerData = 1,
-                                              originalData = TRUE,
-                                              source = "A", ...) {
-    plotly::ggplotly(plot(p, ...), width = width, height = height,
-                     tooltip = tooltip, dynamicTicks = dynamicTicks,
-                     layerData = layerData, originalData = originalData,
-                     source = source)
+#' @export
+plotHover.ArrayTimeBySpeciesBySize <- function(x, ...) {
+    plotHover(plot(x, ...), ...)
 }
 
 #' @rdname animate
 #' @usage NULL
 #' @export
 animate.ArrayTimeBySpeciesBySize <- function(x, species = NULL,
-                                             time_range = NULL,
                                              log_x = TRUE,
+                                             log_y = TRUE,
                                              log = NULL,
                                              wlim = c(NA, NA),
                                              llim = c(NA, NA),
                                              ylim = c(NA, NA),
-                                             log_y = TRUE,
+                                             tlim = c(NA, NA),
                                              size_axis = c("w", "l"),
                                              total = FALSE,
                                              background = TRUE,
@@ -324,14 +314,13 @@ animate.ArrayTimeBySpeciesBySize <- function(x, species = NULL,
 
     times <- as.numeric(dimnames(x)[[1]])
     arr <- unclass(x)
-    if (!is.null(time_range)) {
-        if (length(time_range) == 2 && !all(time_range %in% times)) {
-            keep <- times >= time_range[1] & times <= time_range[2]
-        } else {
-            keep <- times %in% time_range
-        }
-        arr <- arr[keep, , , drop = FALSE]
-        times <- times[keep]
+    if (!is.na(tlim[1])) {
+        arr <- arr[times >= tlim[1], , , drop = FALSE]
+        times <- times[times >= tlim[1]]
+    }
+    if (!is.na(tlim[2])) {
+        arr <- arr[times <= tlim[2], , , drop = FALSE]
+        times <- times[times <= tlim[2]]
     }
 
     w <- as.numeric(dimnames(arr)[[3]])